@@ -2,7 +2,7 @@ import * as THREE from "three";
 import { Circle } from "../../../lib/drawing/shape/2d/circle";
 
 export default function CircleShapeElement({ circle }: { circle: Circle }) {
-  let lineGeometry = new THREE.BufferGeometry().setFromPoints(
+  const lineGeometry = new THREE.BufferGeometry().setFromPoints(
     generateCircleLineStrip(circle),
   );
 
@@ -20,8 +20,8 @@ function generateCircleLineStrip(circle: Circle): THREE.Vector3[] {
     return [];
   }
 
-  let points = new Array<THREE.Vector3>(circle.resolution);
-  let angleStep = (circle.circleEnd - circle.circleStart) / points.length;
+  const points = new Array<THREE.Vector3>(circle.resolution);
+  const angleStep = (circle.circleEnd - circle.circleStart) / points.length;
   for (let pIndex = 0; pIndex < circle.resolution + 1; pIndex++) {
     points[pIndex] = new THREE.Vector3(
       circle.radius * Math.cos(circle.circleStart + angleStep * pIndex),
@@ -32,8 +32,8 @@ function generateCircleLineStrip(circle: Circle): THREE.Vector3[] {
 
   // center lines
   // do not draw center lines when circle is closed
-  let clampedStart = circle.circleStart % (Math.PI * 2);
-  let clampedEnd = circle.circleEnd % (Math.PI * 2);
+  const clampedStart = circle.circleStart % (Math.PI * 2);
+  const clampedEnd = circle.circleEnd % (Math.PI * 2);
   if (Math.abs(clampedEnd - clampedStart) > 0.001) {
     points.push(circle.position);
   }
