@@ -1,22 +1,17 @@
-import { extend, ReactThreeFiber } from "@react-three/fiber";
 import * as THREE from "three";
-import { Line } from "../../../lib/drawing/shape/2d/line";
+import { CatmullRomSpline } from "../../../../lib/drawing/shape/curve/catmull_rom_spline";
 
-extend({ Line_: THREE.Line });
-
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      line_: ReactThreeFiber.Object3DNode<THREE.Line, typeof THREE.Line>;
-    }
-  }
-}
-
-export default function LineShapeElement({ line }: { line: Line }) {
-  let geometry = new THREE.BufferGeometry().setFromPoints([line.a, line.b]);
-
-  if (line.properties.dashedProperties) {
-    let dashProps = line.properties.dashedProperties;
+export default function CatmullRomSplineElement({
+  spline,
+}: {
+  spline: CatmullRomSpline;
+}) {
+  const geometry = new THREE.BufferGeometry().setFromPoints(
+    spline.getCurvePoints(),
+  );
+  console.log(spline.getBoundingBox());
+  if (spline.properties.dashedProperties) {
+    const dashProps = spline.properties.dashedProperties;
     return (
       <group>
         <line_
