@@ -1,17 +1,22 @@
-import { CatmullRomSpline } from "../../../../lib/drawing/shape/curve/catmull_rom_spline";
+import { extend, ReactThreeFiber } from "@react-three/fiber";
 import * as THREE from "three";
+import { Line } from "../../../lib/drawing/shape/2d/line";
 
-export default function CatmullRomSplineElement({
-  spline,
-}: {
-  spline: CatmullRomSpline;
-}) {
-  let geometry = new THREE.BufferGeometry().setFromPoints(
-    spline.getCurvePoints(),
-  );
-  console.log(spline.getBoundingBox());
-  if (spline.properties.dashedProperties) {
-    let dashProps = spline.properties.dashedProperties;
+extend({ Line_: THREE.Line });
+
+declare global {
+  namespace JSX {
+    interface IntrinsicElements {
+      line_: ReactThreeFiber.Object3DNode<THREE.Line, typeof THREE.Line>;
+    }
+  }
+}
+
+export default function LineShapeElement({ line }: { line: Line }) {
+  const geometry = new THREE.BufferGeometry().setFromPoints([line.a, line.b]);
+
+  if (line.properties.dashedProperties) {
+    const dashProps = line.properties.dashedProperties;
     return (
       <group>
         <line_
